Update rendering tests and add warning to ColoredTileWithSpacing
A warning has been added to the ColoredTileWithSpacing class in
ColoredTileWithSpacing.cs indicating that the test is broken due
to an incorrect pattern offset.

In RenderingTest.cs, the hashes.Recording property has been
uncommented and set to true.

These changes reflect updates to the expected output of these
rendering tests, likely due to modifications in the rendering
logic or test data.